[C] Implemented ownership checks

+++ src/main/java/com/tasks/business/TasksService.java
@@ -8,8 +8,9 @@ import com.tasks.business.entities.TaskState;
 import com.tasks.business.entities.TaskType;
 import com.tasks.business.entities.User;
 import com.tasks.business.exceptions.DuplicatedResourceException;
-import com.tasks.business.exceptions.InalidStateException;
+import com.tasks.business.exceptions.InvalidStateException;
 import com.tasks.business.exceptions.InstanceNotFoundException;
+import com.tasks.business.exceptions.PermissionException;
 import com.tasks.business.repository.CommentsRepository;
 import com.tasks.business.repository.ProjectsRepository;
 import com.tasks.business.repository.TasksRepository;
@@ -38,8 +39,8 @@ public class TasksService {
     private ProjectsRepository projectsRepository;
     
     @Transactional
-    public Task create(String name, String description, String type, String username, Long projectId) 
-            throws DuplicatedResourceException, InstanceNotFoundException {
+    public Task create(String callerName, String name, String description, String type, String username, Long projectId)
+            throws DuplicatedResourceException, InstanceNotFoundException, PermissionException {
         Optional<User> user = userRepository.findById(username);
         if(!user.isPresent()) {
             throw new UsernameNotFoundException(MessageFormat.format("User {0} does not exist", username));
@@ -52,6 +53,9 @@ public class TasksService {
         if(tasksRepository.findByName(name) != null) {
             throw new DuplicatedResourceException("Task", name, 
                 MessageFormat.format("Task ''{0}'' already exists", name));
+        }
+        if (!project.get().getAdmin().getUsername().equals(callerName)) {
+            throw new PermissionException();
         }
         
         Project projectToUpdate = project.get();
@@ -72,8 +76,8 @@ public class TasksService {
     }
     
     @Transactional
-    public Task update(Long id, String name, String description, String type, String username, Long projectId) 
-        throws InstanceNotFoundException, InalidStateException, DuplicatedResourceException {
+    public Task update(String callerName, Long id, String name, String description, String type, String username, Long projectId)
+            throws InstanceNotFoundException, InvalidStateException, DuplicatedResourceException, PermissionException {
         Optional<User> user = userRepository.findById(username);
         if(!user.isPresent()) {
             throw new UsernameNotFoundException(MessageFormat.format("User {0} does not exist", username));
@@ -83,9 +87,15 @@ public class TasksService {
             throw new InstanceNotFoundException(projectId, "Project", 
                     MessageFormat.format("Project {0} does not exist", projectId));
         }
+        if (!project.get().getAdmin().getUsername().equals(callerName)) {
+            throw new PermissionException();
+        }
         Optional<Task> task = tasksRepository.findById(id);
         if(!task.isPresent()) {
             throw new InstanceNotFoundException(id, "Task" , MessageFormat.format("Task {0} does not exist", id));
+        }
+        if (!task.get().getProject().getAdmin().getUsername().equals(callerName)) {
+            throw new PermissionException();
         }
         Task anotherTask = tasksRepository.findByName(name);
         if(anotherTask != null && !Objects.equals(anotherTask.getTaskId(), id)) {
@@ -94,7 +104,7 @@ public class TasksService {
         }
         Task taskToUpdate = task.get();
         if(taskToUpdate.getState().equals(TaskState.CLOSED)) {
-            throw new InalidStateException("Task " + taskToUpdate.getName() + " is closed.");
+            throw new InvalidStateException("Task " + taskToUpdate.getName() + " is closed.");
         }
         Project prevProject = taskToUpdate.getProject();
         if(!prevProject.getProjectId().equals(projectId)) {
@@ -115,46 +125,55 @@ public class TasksService {
     }    
 
     @Transactional
-    public Task changeResolution(Long id, TaskResolution resolution) 
-            throws InstanceNotFoundException, InalidStateException {
+    public Task changeResolution(String callerName, Long id, TaskResolution resolution)
+            throws InstanceNotFoundException, InvalidStateException, PermissionException {
         Optional<Task> optTask = tasksRepository.findById(id);
         if(!optTask.isPresent()) {
             throw new InstanceNotFoundException(id, "Task" , MessageFormat.format("Task {0} does not exist", id));
         }
         Task task = optTask.get();
+        if (!optTask.get().getOwner().getUsername().equals(callerName)) {
+            throw new PermissionException();
+        }
         if(task.getState().equals(TaskState.CLOSED)) {
-            throw new InalidStateException("Task " + task.getName() + " is closed.");
+            throw new InvalidStateException("Task " + task.getName() + " is closed.");
         }
         task.setResolution(resolution);
         return tasksRepository.save(task);
     }
     
     @Transactional
-    public Task changeProgress(Long id, byte progress) 
-            throws InstanceNotFoundException, InalidStateException {
+    public Task changeProgress(String callerName, Long id, byte progress)
+            throws InstanceNotFoundException, InvalidStateException, PermissionException {
         Optional<Task> optTask = tasksRepository.findById(id);
         if(!optTask.isPresent()) {
             throw new InstanceNotFoundException(id, "Task" , MessageFormat.format("Task {0} does not exist", id));
         }
         Task task = optTask.get();
+        if (!optTask.get().getOwner().getUsername().equals(callerName)) {
+            throw new PermissionException();
+        }
         if(task.getState().equals(TaskState.CLOSED)) {
-            throw new InalidStateException("Task " + task.getName() + " is closed.");
+            throw new InvalidStateException("Task " + task.getName() + " is closed.");
         }
         if(!task.getResolution().equals(TaskResolution.IN_PROGRESS)) {
-            throw new InalidStateException("Task " + task.getName() + " is not in progress.");
+            throw new InvalidStateException("Task " + task.getName() + " is not in progress.");
         }
         task.setProgress(progress);
         return tasksRepository.save(task);
     }
     
     @Transactional
-    public Task changeState(Long id, TaskState state)  throws InstanceNotFoundException {
+    public Task changeState(String callerName, Long id, TaskState state) throws InstanceNotFoundException, PermissionException {
         
         Optional<Task> optTask = tasksRepository.findById(id);
         if(!optTask.isPresent()) {
             throw new InstanceNotFoundException(id, "Task" , MessageFormat.format("Task {0} does not exist", id));
         }
         Task task = optTask.get();
+        if (!task.getProject().getAdmin().getUsername().equals(callerName)) {
+            throw new PermissionException();
+        }
         task.setState(state);
         return tasksRepository.save(task);
     } 
@@ -193,17 +212,20 @@ public class TasksService {
     }
 
     @Transactional()
-    public void removeById(Long id) throws InstanceNotFoundException {
+    public void removeById(String callerName, Long id) throws InstanceNotFoundException, PermissionException {
         Optional<Task> optTask = tasksRepository.findById(id);
         if(!optTask.isPresent()) {
             throw new InstanceNotFoundException(id, "Task" , MessageFormat.format("Task {0} does not exist", id));
+        }
+        if (!optTask.get().getProject().getAdmin().getUsername().equals(callerName)) {
+            throw new PermissionException();
         }
         tasksRepository.delete(optTask.get());
     }
     
     @Transactional
     public Comment addComment(Long taskId, String user, String text) 
-            throws InstanceNotFoundException, InalidStateException {
+            throws InstanceNotFoundException, InvalidStateException {
         Optional<User> optUser = userRepository.findById(user);
         if(!optUser.isPresent()) {
             throw new UsernameNotFoundException(MessageFormat.format("User {0} does not exist", user));
@@ -214,7 +236,7 @@ public class TasksService {
         }
         Task task = optTask.get();
         if(task.getState().equals(TaskState.CLOSED)) {
-            throw new InalidStateException("Task " + task.getName() + " is closed.");
+            throw new InvalidStateException("Task " + task.getName() + " is closed.");
         }
         Comment comment = new Comment();
         comment.setText(text);
